Split view methods out of the Element class.
